\name{ProgPermute-package}
\alias{ProgPermute-package}
\alias{ProgPermute}
\docType{package}
\title{
\packageTitle{ProgPermute}
}
\description{
\packageDescription{ProgPermute}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ProgPermute}
\packageIndices{ProgPermute}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{ProgPermute}

Maintainer: \packageMaintainer{ProgPermute}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
